requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pydantic
rapidfuzz
requests
playwright
